app/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
bcrypt==4.0.1
# Jinja2==3.1.2
sqlalchemy==2.0.10
asyncpg==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.5
# dacite==1.8.1
# fastapi-pagination==0.12.4
requests==2.31.0